Add tests for model fitting utilities
